refacto : corrected all linter errors
