+++ src/hooks/usePOIs.jsx
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import {PointsOfInterestContext} from "../Contexts/PointsOfInterestProvider"
+
+export function usePOIs() {
+  return useContext(PointsOfInterestContext);
+}
